Split DatabaseManager Put config/connection in DatabaseEnvironment and add table prefix Tables/migrations is now SchemaRegistry
